pyproject.toml
[project]
name = "structured-network-autonomy"
version = "0.1.0"
description = "Structured Network Autonomy — governance framework for AI agents in enterprise networks"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "aiosqlite>=0.20.0",
    "alembic>=1.14.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "pyyaml>=6.0.2",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.1",
    "structlog>=24.4.0",
    "slowapi>=0.1.9",
    "aiofiles>=24.1.0",
    "fastmcp>=2.0.0",
    "scrapli[asyncssh]>=2024.7.30",
    "asyncpg>=0.30.0",
    "typer>=0.15.0",
    "bcrypt>=4.2.0",
    "prometheus-client>=0.21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[project.scripts]
sna = "sna.cli:app"

[build-system]
requires = ["setuptools>=75.0.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = ["error"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "S", "B", "A", "C4", "SIM", "TCH"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]
